# docs/openapi/openapi-v2.yaml
openapi: 3.1.0
info:
  title: CircleHaven — Trust Circle API v2 (E2EE X25519/Ed25519)
  version: "2.0.0"
  description: Les corps Auth sont limités à 16 Kio et les corps Messaging à 256 Kio. Les objets d'entrée refusent les propriétés inconnues.
servers:
  - url: https://trust-circle.kavalek.fr
security:
  - bearerAuth: []
    deviceIdHeader: []
    deviceKeyVersionHeader: []
    deviceProofHeader: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token EdDSA/Ed25519 uniquement, audience trust-circle-api et typ=access.
    refreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token uniquement, audience trust-circle-auth et typ=refresh.
    deviceIdHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Id
      description: UUID de l'appareil du compte qui produit la preuve d'accès.
    deviceKeyVersionHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Key-Version
      description: Version entière positive de la clé d'identité Ed25519 de l'appareil.
    deviceProofHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Proof
      description: Signature Ed25519 Base64 canonique de la transcription d'accès de 89 octets liée au jti de l'access token. Les trois en-têtes appareil sont obligatoires ensemble sur les routes Messaging, sauf enrôlement initial.
  schemas:
...
    DeviceRegistrationChallenge:
      type: object
      required: [challengeId, deviceId, challenge, transcript, expiresAt, algorithm]
      properties:
        challengeId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        challenge: { type: string, description: Base64 canonique du nonce CSPRNG de 32 octets }
        transcript: { type: string, description: Base64 des 163 octets exacts à signer }
        expiresAt: { type: string, format: date-time }
        algorithm: { type: string, enum: [Ed25519] }
...
paths:
...
  /api/devices/registrations/challenge:
    post:
      tags: [devices]
      security:
        - bearerAuth: []
      summary: Créer une transcription de preuve Ed25519 à usage unique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deviceId, identityPublicKey, platform, deviceName]
              properties:
                deviceId: { type: string, format: uuid }
                identityPublicKey: { type: string, minLength: 44, maxLength: 44 }
                platform: { type: string, enum: [android, ios, windows, macos, unknown] }
                deviceName: { type: string, minLength: 1, maxLength: 64 }
                bootstrapGrant: { type: string, minLength: 43, maxLength: 43, description: Requis si cette preuve doit réaliser le tout premier bootstrap }
      responses:
        "201":
          description: Challenge créé pour 5 minutes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceRegistrationChallenge' }
        "403": { description: Identifiant d'appareil révoqué ou grant de bootstrap invalide }
        "409": { description: Identité différente déjà liée à cet UUID d'appareil }
        "429": { description: Limite de challenges atteinte }
  /api/devices/registrations/{challengeId}/proof:
    post:
      tags: [devices]
      security:
        - bearerAuth: []
      summary: Prouver la possession de la clé privée d'identité
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [signature]
              properties:
                signature: { type: string, minLength: 88, maxLength: 88, description: Base64 canonique de 64 octets Ed25519 }
      responses:
        "201":
          description: Preuve valide ; premier appareil actif, suivants en attente
          content:
            application/json:
              schema:
                type: object
                required: [deviceId, status, bootstrap]
                properties:
                  deviceId: { type: string, format: uuid }
                  status: { type: string, enum: [pending, active] }
                  bootstrap: { type: boolean }
        "403": { description: "Signature invalide, bootstrap non autorisé ou appareil révoqué" }
        "404": { description: Challenge absent pour le sujet }
        "409": { description: Challenge consommé ou identité en conflit }
        "410": { description: Challenge expiré }
...